+++ index.html
@@ -232,5 +232,5 @@
 
 </body>
 <!-- Подключение скриптов -->
-<script src="js/main.js"></script>
+<script type="module" src="js/main.js"></script>
 </html>
